requirements.txt
pyyaml>=6.0
docker>=6.0.0
textual>=0.41.0
rich>=13.0.0
psutil>=5.9.0
requests>=2.28.0
python-dateutil>=2.8.0